Move files around, minor cleanup
