+++ test/test_bit_array.rb
